Refactoring code for issue 7342, repeated code to produce rich html is collected and placed in a common file
